% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{default_db_ref}
\alias{default_db_ref}
\title{Defines valid taxonomic databases}
\usage{
default_db_ref(defaults = list())
}
\description{
Defines valid taxonomic databases
}
\section{Attribution}{


This code is copied from the code handling options in the knitr package.
}

\keyword{internal}
